fix(workspace): fix default viewer
