+++ packages/portal/feature-audit/viewer/src/lib/viewer.component.ts
@@ -1,4 +1,4 @@
-import { Component, computed, effect, inject, input, model } from '@angular/core';
+import { Component, inject, input } from '@angular/core';
 import { FormControl, FormGroup, ReactiveFormsModule } from '@angular/forms';
 import { MatCard, MatCardContent, MatCardHeader, MatCardTitle } from '@angular/material/card';
 import { MatError, MatFormField, MatLabel } from '@angular/material/form-field';
@@ -6,15 +6,11 @@ import { MatFabButton } from '@angular/material/button';
 import { MatInput } from '@angular/material/input';
 import { RxIf } from '@rx-angular/template/if';
 import { Router } from '@angular/router';
-import { toObservable, toSignal } from '@angular/core/rxjs-interop';
-import { FlowResult } from 'lighthouse';
-import { filter, map, Observable, switchMap } from 'rxjs';
-import { HttpClient } from '@angular/common/http';
 import { MatProgressSpinner } from '@angular/material/progress-spinner';
 import { NgIf } from '@angular/common';
 import { AuditViewerContainer } from './container/audit-viewer.container';
 import { FlowResultComponent } from './flow-result.component';
-import { AuditSummary, AuditSummaryComponent } from '@app-speed/portal-ui/audit-summary';
+import { AuditSummaryComponent } from '@app-speed/portal-ui/audit-summary';
 import { ViewerStepDetailComponent } from './viewer-container/viewer-step-details.component';
 
 @Component({
@@ -44,11 +40,8 @@ import { ViewerStepDetailComponent } from './viewer-container/viewer-step-detail
       </mat-card>
     </form>
 
-    @if (auditSummary(); as summary) {
-      <ui-audit-summary [(activeIndex)]="activeIndex" [auditSummary]="summary" />
-    }
-    @if (auditStep(); as step) {
-      <viewer-step-detail [stepDetails]="step" />
+    @if (lookupForm.disabled && lookupForm.controls.key.getRawValue(); as auditId) {
+      <viewer-container [auditId]="auditId" />
     }
   `,
   styles: `
@@ -92,43 +85,10 @@ import { ViewerStepDetailComponent } from './viewer-container/viewer-step-detail
 })
 export class ViewerComponent {
   auditId = input<string>('');
-  // 2024-08-18T05_160t0WjP64yCyRK0xVadug
+
   readonly #router = inject(Router);
-  readonly #api = inject(HttpClient);
 
-  flowResult$: Observable<FlowResult> = toObservable(this.auditId).pipe(
-    filter((audit) => audit !== undefined),
-    filter((audit) => audit !== ''),
-    map((auditKey: string) => `https://deepblue-userflow-records.s3.eu-central-1.amazonaws.com/${auditKey}.uf.json`),
-    switchMap((auditKey) => this.#api.get<FlowResult>(auditKey)),
-  );
-
-  auditSummary = toSignal<AuditSummary>(
-    this.flowResult$.pipe(
-      filter(Boolean),
-      map(({ steps }) => {
-        return steps.map(({ lhr: { fullPageScreenshot, categories, gatherMode }, name }) => ({
-          screenShot: fullPageScreenshot?.screenshot.data || '',
-          title: name,
-          subTitle: gatherMode,
-          categoryScores: Object.values(categories).map(({ title, score }) => ({
-            name: title,
-            score: parseInt(((score || 0) * 100).toFixed(0), 10),
-          })),
-        }));
-      }),
-    ),
-  );
-  results = toSignal(this.flowResult$.pipe(filter(Boolean)));
-  auditStep = computed(() => {
-    const results = this.results();
-    const activeStep = this.activeIndex();
-    if (!results || activeStep === undefined) {
-      return;
-    }
-    return results.steps[activeStep];
-  });
-
+  // 2024-08-18T05_160t0WjP64yCyRK0xVadug
   public readonly lookupForm = new FormGroup({
     key: new FormControl<string>('2024-10-02T15_12XDZW3hNNpiK3hvyp7FLM'),
   });
@@ -137,13 +97,6 @@ export class ViewerComponent {
     this.lookupForm.disable();
     this.#router.navigate([], {
       queryParams: { auditId: this.lookupForm.controls.key.value },
-    });
-  }
-
-  activeIndex = model<number>(0);
-  constructor() {
-    effect(() => {
-      console.log('WOLOLO', this.activeIndex());
     });
   }
 }
